Fixed lots of bugs during recent live recording Added VGA2USB device Allowed SimpleProcessor to support format change during processing

+++ trunk/ViCoVRE/ViCoMedia/src/main/java/com/googlecode/vicovre/media/protocol/civil/LiveStream.java
@@ -84,7 +84,7 @@ public class LiveStream implements PushBufferStream, CaptureObserver,
 
     private com.lti.civil.VideoFormat setFormat = null;
 
-    private int sequence = 0;
+    private long sequence = 0;
 
     public LiveStream() throws CaptureException {
         captureSystem =
